apps/backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
langchain==1.0.8
langchain-aws==1.0.0
langgraph==1.0.3
pydantic==2.12.4
python-dotenv==1.2.1
boto3==1.41.1
typing-extensions==4.15.0
PyPDF2==3.0.1
langchain-community==0.4.1
tavily-python==0.7.13
langsmith==0.4.46
firecrawl-py==4.8.0
